requirements.txt
setuptools
streamlit
pandas
Pillow
torch
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
xlsxwriter